# openapi.yaml
openapi: 3.0.3
info:
  title: Down Dirty 84 AI Tuning MVP API
  version: 0.3.0
servers:
- url: http://localhost:8080
paths:
...
  /api/v1/jobs/{jobId}/diffsets/generate:
    post:
      summary: Generate a diffset for a job run
      parameters:
      - in: path
        name: jobId
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiffSetGenerateRequest'
      responses:
        '200':
          description: DiffSet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiffSet'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/jobs/{jobId}/diffsets:
    get:
      summary: List diffsets for a job run
      parameters:
      - in: path
        name: jobId
        required: true
        schema:
          type: string
          format: uuid
      - in: query
        name: runId
        required: false
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: DiffSet summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DiffSetSummary'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/diffsets/{diffSetId}:
    get:
      summary: Get diffset by ID
      parameters:
      - in: path
        name: diffSetId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: DiffSet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiffSet'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/diffsets/{diffSetId}/export/summary:
    post:
      summary: Export branded summary text for diffset
      parameters:
      - in: path
        name: diffSetId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Summary text
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportSummaryResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/diffsets/{diffSetId}/export/csv:
    post:
      summary: Export CSV changelist for diffset
      parameters:
      - in: path
        name: diffSetId
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportCsvRequest'
      responses:
        '200':
          description: CSV
          content:
            text/csv:
              schema:
                type: string
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/start:
    post:
      summary: Start magic link login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthStartRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthStartResponse'
  /api/v1/auth/verify:
    post:
      summary: Verify magic link token and create session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthVerifyRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthVerifyResponse'
        '401':
          description: Unauthorized
...
  /api/v1/jobs:
    get:
      summary: List jobs
      responses:
        '200':
          description: OK
    post:
      summary: Create job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJobRequest'
      responses:
        '201':
          description: Created
  /api/v1/jobs/{jobId}:
    get:
      summary: Get job
      parameters:
      - name: jobId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
        '404':
          description: Not found
...
  /api/v1/admin/jobs:
    get:
      summary: Admin list jobs
      responses:
        '200':
          description: OK
        '403':
          description: Forbidden
  /api/v1/admin/jobs/{jobId}:
    get:
      summary: Admin get job
      responses:
        '200':
          description: OK
        '403':
          description: Forbidden
    patch:
      summary: Admin update job
      responses:
        '200':
          description: OK
        '403':
          description: Forbidden
...
components:
  schemas:
    ErrorResponse:
      type: object
      required:
      - error
      - message
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    DiffSetGenerateRequest:
      type: object
      required:
      - runId
      - generator
      properties:
        runId:
          type: string
          format: uuid
        generator:
          type: string
          enum:
          - GM_LS_MAF_V1
        options:
          type: object
          additionalProperties: true
    DiffSetSummary:
      type: object
      required:
      - diffSetId
      - runId
      - name
      - createdAt
      - itemCounts
      properties:
        diffSetId:
          type: string
          format: uuid
        runId:
          type: string
          format: uuid
        name:
          type: string
        source:
          type:
          - string
          - 'null'
        createdAt:
          type:
          - string
          - 'null'
          format: date-time
        itemCounts:
          type: object
          required:
          - total
          - approved
          - suggested
          - rejected
          properties:
            total:
              type: integer
            approved:
              type: integer
            suggested:
              type: integer
            rejected:
              type: integer
    DiffSet:
      type: object
      required:
      - diffSetId
      - name
      - items
      properties:
        diffSetId:
          type: string
          format: uuid
        jobId:
          type:
          - string
          - 'null'
        runId:
          type:
          - string
          - 'null'
        name:
          type: string
        source:
          type:
          - string
          - 'null'
        createdAt:
          type:
          - string
          - 'null'
          format: date-time
        items:
          type: array
          items:
            $ref: '#/components/schemas/DiffItem'
        diagnostics:
          type: object
          additionalProperties: true
    DiffItem:
      type: object
      required:
      - itemId
      - type
      - path
      - coordinates
      - valueMode
      - units
      - after
      - status
      properties:
        itemId:
          type: integer
        type:
          type: string
          enum:
          - CURVE_POINT
        path:
          type: string
        coordinates:
          type: object
          required:
          - x
          properties:
            x:
              type: number
            y:
              type:
              - number
              - 'null'
        valueMode:
          type: string
          enum:
          - MULTIPLIER
        before: {}
        after:
          type: number
        units:
          type: string
          enum:
          - multiplier
        confidence:
          type: number
          minimum: 0
          maximum: 1
        reason:
          type: string
        applyNotes:
          type: string
        status:
          type: string
          enum:
          - APPROVED
          - SUGGESTED
          - REJECTED
          - APPROVED_OVERRIDE
        tags:
          type: array
          items:
            type: string
    ExportSummaryResponse:
      type: object
      required:
      - text
      properties:
        text:
          type: string
    ExportCsvRequest:
      type: object
      properties:
        includeSuggested:
          type: boolean
          default: false
        minConfidence:
          type: number
          minimum: 0
          maximum: 1
          default: 0.45
    User:
      type: object
      required:
      - id
      - email
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
    AuthStartRequest:
      type: object
      required:
      - email
      properties:
        email:
          type: string
    AuthStartResponse:
      type: object
      required:
      - ok
      properties:
        ok:
          type: boolean
    AuthVerifyRequest:
      type: object
      required:
      - token
      properties:
        token:
          type: string
    AuthVerifyResponse:
      type: object
      required:
      - user
      properties:
        user:
          $ref: '#/components/schemas/User'
...
    CreateJobRequest:
      type: object
      required:
      - serviceType
      - platform
      properties:
        serviceType:
          type: string
        platform:
          type: string
        engineFamily:
          type: string
        vehicle:
          type: string
        ecu:
          type: string
        notes:
          type: string
...
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: dd84_session